[TUNA]: Add a "What is this?" menu entry
Right now its just for kernel threads, but it should be for any
component, and perhaps activated using an icon that when pointed to
something brings up a popup box with the explanation.

This is mostly to cope with older versions of pygtk2 where tooltips
for tree store entries aren't implemented.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>