oscilloscope: use io_add_watch instead of timeout_add
I have to read the GTK documentation
I have to read the GTK documentation
I have to read the GTK documentation
.
.
.

Now the update on the screen is as fast as your computer and metric source, so
use with care.

We may want to have an artificial interval added, so using timeout_add
probably will be used when, for instance, replaying a file with metrics as in:

cyclictest -v > latencies_when_doing_make-j255
oscilloscope --pace 2 < latencies_when_doing_make-j255

So that we can have, say, 2ms of delay from one sample to the next.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>