tuna: Ignore rtprio when changing sched policy to SCHED_OTHER
When the user changes the sched policy to SCHED_OTHER the rtprio entry is
grayed, later, when thread_set_attributes is called it was trying to set
SCHED_OTHER with the previous, SCHED_RR/FIFO rtprio.

That, of course doesn't works, so fix it by setting the rtprio to 0 when
changing to SCHED_OTHER.

Thanks to Jon Masters for pointing out this annoying problem that required that
first the user set the rtprio to 0, then the policy to OTHER and then hit OK to
have it working.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>