tuna-cmd: Introduce --sockets/-S
To specify a CPU socket to be isolated, included or to move specific threads to
it.

To isolate a cpu socket:

tuna --sockets 0 --isolate

Use --include to include the cores in the socket into the affinity mask of all
threads (except kernel threads).

To set the affinity mask of threads:

tuna --sockets 1 --threads 1,4,200 --move

If the children threads are to be moved too use:

tuna --affect_children --sockets 1 --threads 1,4,200 --move

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>