gui: procview: update_rows should do a paranoia check on dup rows
In procview the rows are used while checking if new threads appeared and if old
ones died, but somehow with some pygtk2 releases we could end up with
duplicates, i.e. a thread that already was in the list_store would be added
again because the traversal of the existing rows wouldn't find it even with it
being there.

Add a workaround to remove the previous entry if it has the same pid.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>