tuna: Catch OSError exceptions from python-schedutils
The python-schedutils package incorrectly raise SystemError instead of
OSError when a specified PID did not exist. It has since been corrected,
but tuna is still looking for the SystemError exception everywhere.

Look for both, so that tuna works with old and new versions, a followup
patch will tighten the handling, checking if the error is '3', no such
process, if not, re-raise the exception.

Signed-off-by: Guy Streeter <streeter@redhat.com>
[ committer note: Don't replace SystemError with OSError, handle both ]
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>